+++ python/GafferTest/TransformPlugTest.py
@@ -161,5 +161,43 @@ class TransformPlugTest( GafferTest.TestCase ) :
 		self.assertEqual( s2["n"]["p"]["translate"].getValue(), imath.V3f( 1, 2, 3 ) )
 		self.assertEqual( s2["n"]["p"]["rotate"].getValue(), imath.V3f( 4, 5, 6 ) )
 
+	def testDefaultValues( self ) :
+
+		s = Gaffer.ScriptNode()
+		s["n"] = Gaffer.Node()
+		s["n"]["user"]["p"] = Gaffer.TransformPlug(
+			defaultTranslate = imath.V3f( 1, 2, 3 ),
+			defaultRotate = imath.V3f( 4, 5, 6 ),
+			defaultScale = imath.V3f( 7, 8, 9 ),
+			defaultPivot = imath.V3f( 10, 11, 12 ),
+			flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic
+		)
+		self.assertEqual( s["n"]["user"]["p"]["translate"].defaultValue(), imath.V3f( 1, 2, 3 ) )
+		self.assertEqual( s["n"]["user"]["p"]["rotate"].defaultValue(), imath.V3f( 4, 5, 6 ) )
+		self.assertEqual( s["n"]["user"]["p"]["scale"].defaultValue(), imath.V3f( 7, 8, 9 ) )
+		self.assertEqual( s["n"]["user"]["p"]["pivot"].defaultValue(), imath.V3f( 10, 11, 12 ) )
+
+		s["n"]["user"]["p2"] = s["n"]["user"]["p"].createCounterpart( "p2", Gaffer.Plug.Direction.In )
+		self.assertEqual( s["n"]["user"]["p2"]["translate"].defaultValue(), imath.V3f( 1, 2, 3 ) )
+		self.assertEqual( s["n"]["user"]["p2"]["rotate"].defaultValue(), imath.V3f( 4, 5, 6 ) )
+		self.assertEqual( s["n"]["user"]["p2"]["scale"].defaultValue(), imath.V3f( 7, 8, 9 ) )
+		self.assertEqual( s["n"]["user"]["p2"]["pivot"].defaultValue(), imath.V3f( 10, 11, 12 ) )
+
+		s["n"]["user"]["p2"]["translate"].setValue( imath.V3f( -1, -2, -3 ) )
+		s["n"]["user"]["p2"]["translate"].resetDefault()
+		self.assertEqual( s["n"]["user"]["p2"]["translate"].defaultValue(), imath.V3f( -1, -2, -3 ) )
+
+		s2 = Gaffer.ScriptNode()
+		s2.execute( s.serialise() )
+
+		self.assertEqual( s2["n"]["user"]["p"]["translate"].defaultValue(), imath.V3f( 1, 2, 3 ) )
+		self.assertEqual( s2["n"]["user"]["p"]["rotate"].defaultValue(), imath.V3f( 4, 5, 6 ) )
+		self.assertEqual( s2["n"]["user"]["p"]["scale"].defaultValue(), imath.V3f( 7, 8, 9 ) )
+		self.assertEqual( s2["n"]["user"]["p"]["pivot"].defaultValue(), imath.V3f( 10, 11, 12 ) )
+		self.assertEqual( s2["n"]["user"]["p2"]["translate"].defaultValue(), imath.V3f( -1, -2, -3 ) )
+		self.assertEqual( s2["n"]["user"]["p2"]["rotate"].defaultValue(), imath.V3f( 4, 5, 6 ) )
+		self.assertEqual( s2["n"]["user"]["p2"]["scale"].defaultValue(), imath.V3f( 7, 8, 9 ) )
+		self.assertEqual( s2["n"]["user"]["p2"]["pivot"].defaultValue(), imath.V3f( 10, 11, 12 ) )
+
 if __name__ == "__main__":
 	unittest.main()