+++ Changes.md
@@ -10,6 +10,17 @@ Fixes
 -----
 
 - Checkerboard : Checker colors are now exactly equal to the colorA and colorB parameters.  Previously, there were very tiny floating point errors which grew larger as the distance from origin increased.
+- Transform2DPlug : Fixed serialisation of dynamic plugs, such as plugs promoted to a Box.
+
+API
+---
+
+- TransformPlug/Transform2DPlug : Added constructor arguments for specifying child plug default values.
+
+Breaking Changes
+----------------
+
+- TransformPlug/Transform2DPlug : Added constructor arguments.
 
 0.59.0.0b3 (relative to 0.59.0.0b2)
 ==========